+++ crates/polars-plan/src/dsl/function_expr/range/linear_space.rs
@@ -2,7 +2,7 @@ use arrow::temporal_conversions::MILLISECONDS_IN_DAY;
 use polars_core::prelude::*;
 use polars_ops::series::{new_linear_space_f32, new_linear_space_f64, ClosedInterval};
 
-use super::utils::{ensure_range_bounds_contain_exactly_one_value, linear_spaces_impl_broadcast};
+use super::utils::{build_nulls, ensure_range_bounds_contain_exactly_one_value};
 
 const CAPACITY_FACTOR: usize = 5;
 
@@ -210,4 +210,157 @@ pub(super) fn linear_spaces(
             out.cast(&to_type)
         },
     }
+}
+
+/// Create a ranges column from the given start/end columns and a range function.
+pub(super) fn linear_spaces_impl_broadcast<T, F>(
+    start: &ChunkedArray<T>,
+    end: &ChunkedArray<T>,
+    num_samples: &UInt64Chunked,
+    linear_space_impl: F,
+    builder: &mut ListPrimitiveChunkedBuilder<T>,
+) -> PolarsResult<Column>
+where
+    T: PolarsFloatType,
+    F: Fn(T::Native, T::Native, u64, &mut ListPrimitiveChunkedBuilder<T>) -> PolarsResult<()>,
+    ListPrimitiveChunkedBuilder<T>: ListBuilderTrait,
+{
+    match (start.len(), end.len(), num_samples.len()) {
+        (len_start, len_end, len_samples) if len_start == len_end && len_start == len_samples => {
+            // (n, n, n)
+            build_linear_spaces::<_, _, _, T, F>(
+                start.iter(),
+                end.iter(),
+                num_samples.iter(),
+                linear_space_impl,
+                builder,
+            )?;
+        },
+        // (1, n, n)
+        (1, len_end, len_samples) if len_end == len_samples => {
+            let start_value = start.get(0);
+            if start_value.is_some() {
+                build_linear_spaces::<_, _, _, T, F>(
+                    std::iter::repeat(start_value),
+                    end.iter(),
+                    num_samples.iter(),
+                    linear_space_impl,
+                    builder,
+                )?
+            } else {
+                build_nulls(builder, len_end)
+            }
+        },
+        // (n, 1, n)
+        (len_start, 1, len_samples) if len_start == len_samples => {
+            let end_value = end.get(0);
+            if end_value.is_some() {
+                build_linear_spaces::<_, _, _, T, F>(
+                    start.iter(),
+                    std::iter::repeat(end_value),
+                    num_samples.iter(),
+                    linear_space_impl,
+                    builder,
+                )?
+            } else {
+                build_nulls(builder, len_start)
+            }
+        },
+        // (n, n, 1)
+        (len_start, len_end, 1) if len_start == len_end => {
+            let num_samples_value = num_samples.get(0);
+            if num_samples_value.is_some() {
+                build_linear_spaces::<_, _, _, T, F>(
+                    start.iter(),
+                    end.iter(),
+                    std::iter::repeat(num_samples_value),
+                    linear_space_impl,
+                    builder,
+                )?
+            } else {
+                build_nulls(builder, len_start)
+            }
+        },
+        // (n, 1, 1)
+        (len_start, 1, 1) => {
+            let end_value = end.get(0);
+            let num_samples_value = num_samples.get(0);
+            match (end_value, num_samples_value) {
+                (Some(_), Some(_)) => build_linear_spaces::<_, _, _, T, F>(
+                    start.iter(),
+                    std::iter::repeat(end_value),
+                    std::iter::repeat(num_samples_value),
+                    linear_space_impl,
+                    builder,
+                )?,
+                _ => build_nulls(builder, len_start),
+            }
+        },
+        // (1, n, 1)
+        (1, len_end, 1) => {
+            let start_value = start.get(0);
+            let num_samples_value = num_samples.get(0);
+            match (start_value, num_samples_value) {
+                (Some(_), Some(_)) => build_linear_spaces::<_, _, _, T, F>(
+                    std::iter::repeat(start_value),
+                    end.iter(),
+                    std::iter::repeat(num_samples_value),
+                    linear_space_impl,
+                    builder,
+                )?,
+                _ => build_nulls(builder, len_end),
+            }
+        },
+        // (1, 1, n)
+        (1, 1, len_num_samples) => {
+            let start_value = start.get(0);
+            let end_value = end.get(0);
+            match (start_value, end_value) {
+                (Some(_), Some(_)) => build_linear_spaces::<_, _, _, T, F>(
+                    std::iter::repeat(start_value),
+                    std::iter::repeat(end_value),
+                    num_samples.iter(),
+                    linear_space_impl,
+                    builder,
+                )?,
+                _ => build_nulls(builder, len_num_samples),
+            }
+        },
+        (len_start, len_end, len_num_samples) => {
+            polars_bail!(
+                ComputeError:
+                "lengths of `start` ({}), `end` ({}), and `num_samples` ({}) do not match",
+                len_start, len_end, len_num_samples
+            )
+        },
+    };
+    let out = builder.finish().into_column();
+    Ok(out)
+}
+
+/// Iterate over a start and end column and create a range for each entry.
+fn build_linear_spaces<I, J, K, T, F>(
+    start: I,
+    end: J,
+    num_samples: K,
+    linear_space_impl: F,
+    builder: &mut ListPrimitiveChunkedBuilder<T>,
+) -> PolarsResult<()>
+where
+    I: Iterator<Item = Option<T::Native>>,
+    J: Iterator<Item = Option<T::Native>>,
+    K: Iterator<Item = Option<u64>>,
+    T: PolarsFloatType,
+    F: Fn(T::Native, T::Native, u64, &mut ListPrimitiveChunkedBuilder<T>) -> PolarsResult<()>,
+    ListPrimitiveChunkedBuilder<T>: ListBuilderTrait,
+{
+    for ((start, end), num_samples) in start.zip(end).zip(num_samples) {
+        match (start, end, num_samples) {
+            (Some(start), Some(end), Some(num_samples)) => {
+                linear_space_impl(start, end, num_samples, builder)?
+            },
+            _ => builder.append_null(),
+        }
+    }
+    Ok(())
 }
